update int to long long to allow bigger hashes
updating every int type to long long

+++ gss/globals.hpp
@@ -1,1 +1,1 @@
-extern int collisions;
+extern long long collisions;
